Change architecture s.t. property adding problems do not occur
