--- modules/luigi/luigi.py
@@ -1,2 +1,0 @@
-import ravestate as rs
-mod = rs.Module(name="Luigi")
